More hacks to write the correct addends when changing from RELA to REL
When using the external capsizefix command the cap_relocs location would
be the offset relative to the start of the section instead of the load
address. Work around this by calling DynReloc::getAddend after the output
section addresses have been assigned and treating cap_relocs locations as
never preemptible. This does not affect the LLD integrated capsizefix which
should hopefully soon be able to completely replace external capsizefix.

Fixes the failing ELF/cheri/simple-cap-reloc.c test (started when the
RELA->REL hack was added)
hack was