Fix cheribsd build after 36cef549e5a4c857dc4239cebcb7f34b5f3c652e
CheriBSD compiles with --fatal-warnings so don't warn in the expected case.
Commit 36cef5 removed the assertion that the symbol for which we can't find
a symbol starts with .rodata.str, replacing that with a warning. However,
we should not warn if it is against an offset in .rodata.str since clang
will never emit symbols for such objects.