[UPSTREAM?][ELF] Make .eh_frame relro if one of the inputs is SHF_WRITE
Previously .eh_frame was always SHF_ALLOC and read-only. However, for
MIPS64 PIC code clang emits a R_MIPS_64 relocation that will be converted
to a dynamic relocation so .eh_frame can't be read-only in that case.
Fix this by making .eh_frame writable if one of the input sections is has
the SHF_WRITE flag.


This should hopefully fix the need for the -Wl,-z,notext everywhere in
CheriBSD.