Add a new symbol _HAS__DYNAMIC
This can be used by the csu startup code to check if it is dynamically
linked in the purecap ABI. Getting the value of _DYNAMIC without globals
is really tricky and there is no good solution for MIPS since pcrel is
restricted to 32 bits. Now a simple `ori $1, $0, %lo(_HAS__DYNAMIC)` is
enough to perform this check without any loads or 7 instruction `dla`.